combined hackeeg_test.py and hackeeg_test_speed.py for simplicity
- quiet option now does performance test
